Check directory exists before trying to delete it
https://github.com/StaticPatch/StaticPatch/issues/1

+++ staticpatch/staticpatchcore/purge_deleted.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 
 from django.conf import settings
@@ -17,4 +18,5 @@ def purge_deleted():
     for build in BuildModel.objects.filter(deleted_at__isnull=False, deleted_at__lt=cutoff_time):
         dir_name = build.get_full_file_storage_directory()
         build.delete()
-        shutil.rmtree(dir_name)
+        if os.path.isdir(dir_name):
+            shutil.rmtree(dir_name)
